+++ iam-policy-json-to-terraform.go
@@ -3,6 +3,7 @@ package main
 import "fmt"
 import (
 	"bufio"
+	"flag"
 	"github.com/flosell/iam-policy-json-to-terraform/converter"
 	"io/ioutil"
 	"log"
@@ -10,6 +11,8 @@ import (
 )
 
 func main() {
+	policyName := flag.String("name", "policy", "name of the policy in generated hcl")
+	flag.Parse()
 	reader := bufio.NewReader(os.Stdin)
 	b, err := ioutil.ReadAll(reader)
 
@@ -17,7 +20,7 @@ func main() {
 		log.Fatal("unable to read stdin: ", err)
 	}
 
-	converted, err := converter.Convert(b)
+	converted, err := converter.Convert(*policyName, b)
 
 	if err != nil {
 		log.Fatal("unable to convert: ", err)